+++ doc/changelog.md
@@ -3,7 +3,8 @@
 * **0.13.7-SNAPSHOT**
   - Fix default for "cleanup" in build configuration to `true` (as documented) (#338)
   - Fix dynamic host property update in port mapping (#323)  
-  
+  - New goal 'docker:source' for attaching a Docker tar archive to the Maven project with an classifier "docker-<alias>" (#311) 
+
 * **0.13.6**
   - Don't use user from image when pulling base images ([#147](https://github.com/rhuss/docker-maven-plugin/issues/147))
   - Add a new assembly descriptor reference  `hawt-app` for using asseblies created by